codegen: Generate AST structs
