+++ src/data.rs
@@ -1,43 +1,10 @@
 use super::*;
 
 #[allow(unreachable_pub)] // https://github.com/rust-lang/rust/issues/57411
-pub use crate::ast_enum::{Fields, Visibility};
-
-ast_struct! {
-    /// An enum variant.
-    pub struct Variant {
-        /// Attributes tagged on the variant.
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
-        pub(crate) attrs: Vec<Attribute>,
-
-        /// Name of the variant.
-        pub(crate) ident: Ident,
-
-        /// Content stored in the variant.
-        pub(crate) fields: Fields,
-
-        /// Explicit discriminant: `Variant = 1`
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        pub(crate) discriminant: Option<Expr>,
-    }
-}
-
-ast_struct! {
-    /// Named fields of a struct or struct variant such as `Point { x: f64,
-    /// y: f64 }`.
-    #[serde(transparent)]
-    pub struct FieldsNamed {
-        pub(crate) named: Punctuated<Field>,
-    }
-}
-
-ast_struct! {
-    /// Unnamed fields of a tuple struct or tuple variant such as `Some(T)`.
-    #[serde(transparent)]
-    pub struct FieldsUnnamed {
-        pub(crate) unnamed: Punctuated<Field>,
-    }
-}
+pub use crate::{
+    ast_enum::{Fields, Visibility},
+    ast_struct::{FieldsNamed, FieldsUnnamed, Variant, VisRestricted},
+};
 
 impl Fields {
     pub(crate) fn is_named(&self) -> bool {
@@ -84,6 +51,7 @@ ast_struct! {
         #[serde(default, skip_serializing_if = "Option::is_none")]
         pub(crate) ident: Option<Ident>,
 
+        // TODO: can remove?
         #[serde(default, skip_serializing_if = "not")]
         pub(crate) colon_token: bool,
 
@@ -104,15 +72,5 @@ impl Visibility {
 impl Default for Visibility {
     fn default() -> Self {
         Visibility::Inherited
-    }
-}
-
-ast_struct! {
-    /// A visibility level restricted to some path: `pub(self)` or
-    /// `pub(super)` or `pub(crate)` or `pub(in some::module)`.
-    pub struct VisRestricted {
-        #[serde(default, skip_serializing_if = "not")]
-        pub(crate) in_token: bool,
-        pub(crate) path: Box<Path>,
     }
 }
