+++ src/generics.rs
@@ -1,7 +1,12 @@
 use super::*;
 
 #[allow(unreachable_pub)] // https://github.com/rust-lang/rust/issues/57411
-pub use crate::ast_enum::{GenericParam, TraitBoundModifier, TypeParamBound, WherePredicate};
+pub use crate::{
+    ast_enum::{GenericParam, TraitBoundModifier, TypeParamBound, WherePredicate},
+    ast_struct::{
+        BoundLifetimes, ConstParam, PredicateEq, PredicateLifetime, TraitBound, WhereClause,
+    },
+};
 
 ast_struct! {
     /// Lifetimes and type parameters attached to a declaration of a function,
@@ -55,46 +60,8 @@ ast_struct! {
     }
 }
 
-ast_struct! {
-    /// A const generic parameter: `const LENGTH: usize`.
-    pub struct ConstParam {
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
-        pub(crate) attrs: Vec<Attribute>,
-        pub(crate) ident: Ident,
-        pub(crate) ty: Type,
-        #[serde(default, skip_serializing_if = "not")]
-        pub(crate) eq_token: bool,
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        pub(crate) default: Option<Expr>,
-    }
-}
-
-ast_struct! {
-    /// A set of bound lifetimes: `for<'a, 'b, 'c>`.
-    #[derive(Default)]
-    #[serde(transparent)]
-    pub struct BoundLifetimes {
-        pub(crate) lifetimes: Punctuated<LifetimeDef>,
-    }
-}
-
-ast_struct! {
-    /// A trait used as a bound on a type parameter.
-    pub struct TraitBound {
-        #[serde(default, skip_serializing_if = "not")]
-        pub(crate) paren_token: bool,
-        #[serde(default, skip_serializing_if = "TraitBoundModifier::is_none")]
-        pub(crate) modifier: TraitBoundModifier,
-        /// The `for<'a>` in `for<'a> Foo<&'a T>`
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        pub(crate) lifetimes: Option<BoundLifetimes>,
-        /// The `Foo<&'a T>` in `for<'a> Foo<&'a T>`
-        pub(crate) path: Path,
-    }
-}
-
 impl TraitBoundModifier {
-    fn is_none(&self) -> bool {
+    pub(crate) fn is_none(&self) -> bool {
         match self {
             TraitBoundModifier::None => true,
             TraitBoundModifier::Maybe => false,
@@ -109,15 +76,6 @@ impl Default for TraitBoundModifier {
 }
 
 ast_struct! {
-    /// A `where` clause in a definition: `where T: Deserialize<'de>, D:
-    /// 'static`.
-    #[serde(transparent)]
-    pub struct WhereClause {
-        pub(crate) predicates: Punctuated<WherePredicate>,
-    }
-}
-
-ast_struct! {
     /// A type predicate in a `where` clause: `for<'c> Foo<'c>: Trait<'c>`.
     pub struct PredicateType {
         /// Any lifetimes from a `for` binding
@@ -126,23 +84,9 @@ ast_struct! {
         /// The type being bounded
         pub(crate) bounded_ty: Type,
         /// Trait and lifetime bounds (`Clone+Send+'static`)
+        // TODO: should allow default?
+        // #[serde(default, skip_serializing_if = "Punctuated::is_empty")]
         pub(crate) bounds: Punctuated<TypeParamBound>,
-    }
-}
-
-ast_struct! {
-    /// A lifetime predicate in a `where` clause: `'a: 'b + 'c`.
-    pub struct PredicateLifetime {
-        pub(crate) lifetime: Lifetime,
-        pub(crate) bounds: Punctuated<Lifetime>,
-    }
-}
-
-ast_struct! {
-    /// An equality predicate in a `where` clause (unsupported).
-    pub struct PredicateEq {
-        pub(crate) lhs_ty: Type,
-        pub(crate) rhs_ty: Type,
     }
 }
