+++ src/item.rs
@@ -1,47 +1,15 @@
 use super::*;
 
 #[allow(unreachable_pub)] // https://github.com/rust-lang/rust/issues/57411
-pub use crate::ast_enum::{FnArg, ForeignItem, ImplItem, Item, TraitItem, UseTree};
-
-ast_struct! {
-    /// A constant item: `const MAX: u16 = 65535`.
-    pub struct ItemConst {
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
-        pub(crate) attrs: Vec<Attribute>,
-        #[serde(default, skip_serializing_if = "Visibility::is_inherited")]
-        pub(crate) vis: Visibility,
-        pub(crate) ident: Ident,
-        pub(crate) ty: Box<Type>,
-        pub(crate) expr: Box<Expr>,
-    }
-}
-
-ast_struct! {
-    /// An enum definition: `enum Foo<A, B> { A(A), B(B) }`.
-    pub struct ItemEnum {
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
-        pub(crate) attrs: Vec<Attribute>,
-        #[serde(default, skip_serializing_if = "Visibility::is_inherited")]
-        pub(crate) vis: Visibility,
-        pub(crate) ident: Ident,
-        #[serde(default, skip_serializing_if = "Generics::is_none")]
-        pub(crate) generics: Generics,
-        pub(crate) variants: Punctuated<Variant>,
-    }
-}
-
-ast_struct! {
-    /// An `extern crate` item: `extern crate serde`.
-    pub struct ItemExternCrate {
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
-        pub(crate) attrs: Vec<Attribute>,
-        #[serde(default, skip_serializing_if = "Visibility::is_inherited")]
-        pub(crate) vis: Visibility,
-        pub(crate) ident: Ident,
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        pub(crate) rename: Option<Ident>,
-    }
-}
+pub use crate::{
+    ast_enum::{FnArg, ForeignItem, ImplItem, Item, TraitItem, UseTree},
+    ast_struct::{
+        ForeignItemFn, ForeignItemMacro, ForeignItemStatic, ForeignItemType, ImplItemConst,
+        ImplItemMacro, ImplItemType, ItemConst, ItemEnum, ItemExternCrate, ItemForeignMod,
+        ItemMacro, ItemMacro2, ItemStatic, ItemTraitAlias, ItemType, ItemUnion, ItemUse, Signature,
+        TraitItemConst, TraitItemMacro, UseGroup, UseName, UsePath, UseRename,
+    },
+};
 
 ast_struct! {
     /// A free-standing function: `fn process(n: usize) -> Result<()> { ...
@@ -55,16 +23,6 @@ ast_struct! {
         pub(crate) sig: Signature,
         #[serde(rename = "stmts")]
         pub(crate) block: Box<Block>,
-    }
-}
-
-ast_struct! {
-    /// A block of foreign items: `extern "C" { ... }`.
-    pub struct ItemForeignMod {
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
-        pub(crate) attrs: Vec<Attribute>,
-        pub(crate) abi: Abi,
-        pub(crate) items: Vec<ForeignItem>,
     }
 }
 
@@ -93,33 +51,6 @@ ast_struct! {
 }
 
 ast_struct! {
-    /// A macro invocation, which includes `macro_rules!` definitions.
-    pub struct ItemMacro {
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
-        pub(crate) attrs: Vec<Attribute>,
-        /// The `example` in `macro_rules! example { ... }`.
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        pub(crate) ident: Option<Ident>,
-        #[serde(flatten)]
-        pub(crate) mac: Macro,
-        #[serde(default, skip_serializing_if = "not")]
-        pub(crate) semi_token: bool,
-    }
-}
-
-ast_struct! {
-    /// A 2.0-style declarative macro introduced by the `macro` keyword.
-    pub struct ItemMacro2 {
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
-        pub(crate) attrs: Vec<Attribute>,
-        #[serde(default, skip_serializing_if = "Visibility::is_inherited")]
-        pub(crate) vis: Visibility,
-        pub(crate) ident: Ident,
-        pub(crate) rules: TokenStream,
-    }
-}
-
-ast_struct! {
     /// A module or module declaration: `mod m` or `mod m { ... }`.
     pub struct ItemMod {
         #[serde(default, skip_serializing_if = "Vec::is_empty")]
@@ -127,26 +58,12 @@ ast_struct! {
         #[serde(default, skip_serializing_if = "Visibility::is_inherited")]
         pub(crate) vis: Visibility,
         pub(crate) ident: Ident,
+        // TODO: should not skip_serializing_if
         #[serde(default, skip_serializing_if = "Option::is_none")]
         pub(crate) content: Option<Vec<Item>>,
+        // TODO: can remove
         #[serde(default, skip_serializing_if = "not")]
         pub(crate) semi: bool,
-    }
-}
-
-ast_struct! {
-    /// A static item: `static BIKE: Shed = Shed(42)`.
-    pub struct ItemStatic {
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
-        pub(crate) attrs: Vec<Attribute>,
-        #[serde(default, skip_serializing_if = "Visibility::is_inherited")]
-        pub(crate) vis: Visibility,
-        #[serde(rename = "mut")]
-        #[serde(default, skip_serializing_if = "not")]
-        pub(crate) mutability: bool,
-        pub(crate) ident: Ident,
-        pub(crate) ty: Box<Type>,
-        pub(crate) expr: Box<Expr>,
     }
 }
 
@@ -191,155 +108,6 @@ ast_struct! {
 }
 
 ast_struct! {
-    /// A trait alias: `pub trait SharableIterator = Iterator + Sync`.
-    pub struct ItemTraitAlias {
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
-        pub(crate) attrs: Vec<Attribute>,
-        #[serde(default, skip_serializing_if = "Visibility::is_inherited")]
-        pub(crate) vis: Visibility,
-        pub(crate) ident: Ident,
-        #[serde(default, skip_serializing_if = "Generics::is_none")]
-        pub(crate) generics: Generics,
-        pub(crate) bounds: Punctuated<TypeParamBound>,
-    }
-}
-
-ast_struct! {
-    /// A type alias: `type Result<T> = std::result::Result<T, MyError>`.
-    pub struct ItemType {
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
-        pub(crate) attrs: Vec<Attribute>,
-        #[serde(default, skip_serializing_if = "Visibility::is_inherited")]
-        pub(crate) vis: Visibility,
-        pub(crate) ident: Ident,
-        #[serde(default, skip_serializing_if = "Generics::is_none")]
-        pub(crate) generics: Generics,
-        pub(crate) ty: Box<Type>,
-    }
-}
-
-ast_struct! {
-    /// A union definition: `union Foo<A, B> { x: A, y: B }`.
-    pub struct ItemUnion {
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
-        pub(crate) attrs: Vec<Attribute>,
-        #[serde(default, skip_serializing_if = "Visibility::is_inherited")]
-        pub(crate) vis: Visibility,
-        pub(crate) ident: Ident,
-        #[serde(default, skip_serializing_if = "Generics::is_none")]
-        pub(crate) generics: Generics,
-        pub(crate) fields: FieldsNamed,
-    }
-}
-
-ast_struct! {
-    /// A use declaration: `use std::collections::HashMap`.
-    pub struct ItemUse {
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
-        pub(crate) attrs: Vec<Attribute>,
-        #[serde(default, skip_serializing_if = "Visibility::is_inherited")]
-        pub(crate) vis: Visibility,
-        #[serde(default, skip_serializing_if = "not")]
-        pub(crate) leading_colon: bool,
-        pub(crate) tree: UseTree,
-    }
-}
-
-ast_struct! {
-    /// A path prefix of imports in a `use` item: `std::...`.
-    pub struct UsePath {
-        pub(crate) ident: Ident,
-        pub(crate) tree: Box<UseTree>,
-    }
-}
-
-ast_struct! {
-    /// An identifier imported by a `use` item: `HashMap`.
-    #[serde(transparent)]
-    pub struct UseName {
-        pub(crate) ident: Ident,
-    }
-}
-
-ast_struct! {
-    /// An renamed identifier imported by a `use` item: `HashMap as Map`.
-    pub struct UseRename {
-        pub(crate) ident: Ident,
-        pub(crate) rename: Ident,
-    }
-}
-
-ast_struct! {
-    /// A braced group of imports in a `use` item: `{A, B, C}`.
-    #[serde(transparent)]
-    pub struct UseGroup {
-        pub(crate) items: Punctuated<UseTree>,
-    }
-}
-
-ast_struct! {
-    /// A foreign function in an `extern` block.
-    pub struct ForeignItemFn {
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
-        pub(crate) attrs: Vec<Attribute>,
-        #[serde(default, skip_serializing_if = "Visibility::is_inherited")]
-        pub(crate) vis: Visibility,
-        #[serde(flatten)]
-        pub(crate) sig: Signature,
-    }
-}
-
-ast_struct! {
-    /// A foreign static item in an `extern` block: `static ext: u8`.
-    pub struct ForeignItemStatic {
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
-        pub(crate) attrs: Vec<Attribute>,
-        #[serde(default, skip_serializing_if = "Visibility::is_inherited")]
-        pub(crate) vis: Visibility,
-        #[serde(rename = "mut")]
-        #[serde(default, skip_serializing_if = "not")]
-        pub(crate) mutability: bool,
-        pub(crate) ident: Ident,
-        pub(crate) ty: Box<Type>,
-    }
-}
-
-ast_struct! {
-    /// A foreign type in an `extern` block: `type void`.
-    pub struct ForeignItemType {
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
-        pub(crate) attrs: Vec<Attribute>,
-        #[serde(default, skip_serializing_if = "Visibility::is_inherited")]
-        pub(crate) vis: Visibility,
-        pub(crate) ident: Ident,
-    }
-}
-
-ast_struct! {
-    /// A macro invocation within an extern block.
-    pub struct ForeignItemMacro {
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
-        pub(crate) attrs: Vec<Attribute>,
-        #[serde(flatten)]
-        pub(crate) mac: Macro,
-        #[serde(default, skip_serializing_if = "not")]
-        pub(crate) semi_token: bool,
-    }
-}
-
-ast_struct! {
-    /// An associated constant within the definition of a trait.
-    pub struct TraitItemConst {
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
-        pub(crate) attrs: Vec<Attribute>,
-        pub(crate) ident: Ident,
-        pub(crate) ty: Type,
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        pub(crate) default: Option<Expr>,
-    }
-}
-
-ast_struct! {
     /// A trait method within the definition of a trait.
     pub struct TraitItemMethod {
         #[serde(default, skip_serializing_if = "Vec::is_empty")]
@@ -371,34 +139,6 @@ ast_struct! {
 }
 
 ast_struct! {
-    /// A macro invocation within the definition of a trait.
-    pub struct TraitItemMacro {
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
-        pub(crate) attrs: Vec<Attribute>,
-        #[serde(flatten)]
-        pub(crate) mac: Macro,
-        #[serde(default, skip_serializing_if = "not")]
-        pub(crate) semi_token: bool,
-    }
-}
-
-ast_struct! {
-    /// An associated constant within an impl block.
-    pub struct ImplItemConst {
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
-        pub(crate) attrs: Vec<Attribute>,
-        #[serde(default, skip_serializing_if = "Visibility::is_inherited")]
-        pub(crate) vis: Visibility,
-        #[serde(rename = "default")]
-        #[serde(default, skip_serializing_if = "not")]
-        pub(crate) defaultness: bool,
-        pub(crate) ident: Ident,
-        pub(crate) ty: Type,
-        pub(crate) expr: Expr,
-    }
-}
-
-ast_struct! {
     /// A method within an impl block.
     pub struct ImplItemMethod {
         #[serde(default, skip_serializing_if = "Vec::is_empty")]
@@ -412,61 +152,6 @@ ast_struct! {
         pub(crate) sig: Signature,
         #[serde(rename = "stmts")]
         pub(crate) block: Block,
-    }
-}
-
-ast_struct! {
-    /// An associated type within an impl block.
-    pub struct ImplItemType {
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
-        pub(crate) attrs: Vec<Attribute>,
-        #[serde(default, skip_serializing_if = "Visibility::is_inherited")]
-        pub(crate) vis: Visibility,
-        #[serde(rename = "default")]
-        #[serde(default, skip_serializing_if = "not")]
-        pub(crate) defaultness: bool,
-        pub(crate) ident: Ident,
-        #[serde(default, skip_serializing_if = "Generics::is_none")]
-        pub(crate) generics: Generics,
-        pub(crate) ty: Type,
-    }
-}
-
-ast_struct! {
-    /// A macro invocation within an impl block.
-    pub struct ImplItemMacro {
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
-        pub(crate) attrs: Vec<Attribute>,
-        #[serde(flatten)]
-        pub(crate) mac: Macro,
-        #[serde(default, skip_serializing_if = "not")]
-        pub(crate) semi_token: bool,
-    }
-}
-
-ast_struct! {
-    /// A function signature in a trait or implementation: `unsafe fn
-    /// initialize(&self)`.
-    pub struct Signature {
-        #[serde(rename = "const")]
-        #[serde(default, skip_serializing_if = "not")]
-        pub(crate) constness: bool,
-        #[serde(rename = "async")]
-        #[serde(default, skip_serializing_if = "not")]
-        pub(crate) asyncness: bool,
-        #[serde(rename = "unsafe")]
-        #[serde(default, skip_serializing_if = "not")]
-        pub(crate) unsafety: bool,
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        pub(crate) abi: Option<Abi>,
-        pub(crate) ident: Ident,
-        #[serde(default, skip_serializing_if = "Generics::is_none")]
-        pub(crate) generics: Generics,
-        pub(crate) inputs: Punctuated<FnArg>,
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        pub(crate) variadic: Option<Variadic>,
-        #[serde(default)]
-        pub(crate) output: ReturnType,
     }
 }
